Reviewed-on: https://git.sh-edraft.de/sh-edraft.de/sh_cpl/pulls/79
Closes #77

+++ src/cpl_cli/command/remove_service.py
@@ -52,8 +52,9 @@ class RemoveService(CommandABC):
             project_json.write(json.dumps(content, indent=2))
             project_json.close()
 
-    @staticmethod
-    def _remove_sources(path: str):
+    def _remove_sources(self, path: str):
+        if self._is_simulation:
+            return
         shutil.rmtree(path)
 
     def _create_workspace(self, path: str):
